docker/derived_image/cli/requirements.txt
click~=7.1.2
click-aliases==1.0.1
colorama~=0.4.3
config==0.4.2
dataclasses_json==0.5.2
halo~=0.0.29
natsort==3.3.0
pyyaml~=5.3.1
requests==2.22.0
setuptools~=49.2.0
six~=1.15.0
junit_xml==1.8
dataclasses_json==0.5.2
natsort==3.3.0
config==0.4.2
python-terraform==0.10.1
checkov==1.0.765
jsonpickle==1.4.1
netaddr==0.7.19
arnparse==0.0.2
backoff==1.10.0
packaging==20.4
Pygments==2.7.3
semantic_version==2.8.5